Issue #4 recipe pages (#8)
* Moving devices

changing over to my pc
added:
- User options service to deal with bookmarks and settings if any

* Adding PWA

Added manifest.json to start on PWA and moved all icons in the icons folder for easy access later

* Dynamic routing - changing devices

Mainly Implemented dynamic routing running with a few errors.
Changing over to pc - main reason for commit

* Dynamic routing now setup

Dynamic routing for recipe information is now setup for both Pages and Api,
TODO:
- Authentication to API routes
- SEO

* Recipe page data almost showing

Data is now showing on recipe page there are a few errors with certain data not mapping properly

* Recipe page desktop styling mostly done

- Fixed up SEO, Recipe page items
- got most of the styling done for the
  recipe page.
- changed some colors around

* Finished recipe pages except styling

Everything is finished on the recipe page with getting data. styling isn't finished but that can be a future problem as i'm not completely sure how i want things to look.