Dev (#16)
* navbar done

* navbar, hero image, hero text

- Mostly finished navbar - just need to  add interactions
- Made image stretch whole hero with out visible stretching
- Added the text

* Firebase setup

* Firebase config is partly setup

All info is setup just learning how to implement querying with in the project.

* started testing

- added a testing page to test items
- started learning how to set up firebase
- learnt how classes work

* Connection to Firebase

Added the ability to create and read data from firebase

* Auth setup

* Singup Ui almost finished

* Singup UI almost finished

Need to centre the element in the middle of the page

* Auth setup

* Auth

* firebase admin

* firebase auth now working

- firebase auth is now working
- google signin is almost complete

* Auth mostly complete

- Added all sign in methods
not working: Facebook, Github, email and password.
- pages added for signin / signup

* Issue #4 recipe pages (#8)

* Moving devices

changing over to my pc
added:
- User options service to deal with bookmarks and settings if any

* Adding PWA

Added manifest.json to start on PWA and moved all icons in the icons folder for easy access later

* Dynamic routing - changing devices

Mainly Implemented dynamic routing running with a few errors.
Changing over to pc - main reason for commit

* Dynamic routing now setup

Dynamic routing for recipe information is now setup for both Pages and Api,
TODO:
- Authentication to API routes
- SEO

* Recipe page data almost showing

Data is now showing on recipe page there are a few errors with certain data not mapping properly

* Recipe page desktop styling mostly done

- Fixed up SEO, Recipe page items
- got most of the styling done for the
  recipe page.
- changed some colors around

* Finished recipe pages except styling

Everything is finished on the recipe page with getting data. styling isn't finished but that can be a future problem as i'm not completely sure how i want things to look.

* Delete .VSCodeCounter directory

Signed-off-by: Jacob <77374574+Eirfire@users.noreply.github.com>

* Impilment search bar (#15)

* Aglogia installation

added Aglolia to the project:
- moving devices

* Algolia Search implementation Ui

Built in the search component
Designing out the Ui in progress
Building the Dialog component

* Ui

* Building Dialog

trying to make a dialog that works for Algolia, looking to be something like ones that are on tailwindcss.com, algolia.com or socket.io however I can't figure out how to get it working.

also changing devices so that's the reason for this push

* Algolia Search Dialog is working

Search dialog is now implemented, styling is still to complete

* Search Dialog is now syteled

Completed:
- styled search bar
- learnt how Algolia:
  - styles elements
  -  controls elements shown
- bluring background
- Adding keyboard shortcuts

TODO:
- style the Hit component
- make responsive for mobile

* Responsive text, algolia updates

- responsive text
- Algolia hit types
- updated doc in firestore
- updated recipe type

* Update intergrate.yml

Github action for continuous integration in Node

* Hit component is now styles

- Make things responsive
- Make images work

- Styled component to look some what nice
- learnt how to styles the Hit component
- Reorder how Algolia index's items

* Todo lists

* Added docker and finished search box

- fix up dialog on desktop to center in the middle of the page (not urgent)
- Add ability to click outside of dialog to close it

- fixed styling issues
- Added docker files for people
- key binds added for opening and closing dialog

* trying to detect click outside dialog - changing devices

* Update algolia_search_dialog.tsx

Added code that may work to detect if a user clicks outside the element

* Update algolia_search_dialog.tsx

Signed-off-by: Jacob <77374574+Eirfire@users.noreply.github.com>